+++ tests/test_converters/test_flatten.py
@@ -2,19 +2,30 @@
 
 import pytest
 
-from returns.converters import join
+from returns.context import Context
+from returns.converters import flatten
 from returns.io import IO
 from returns.maybe import Nothing, Some
 from returns.result import Failure, Success
 
 
 @pytest.mark.parametrize(('container', 'merged'), [
+    # Flattens:
     (IO(IO(1)), IO(1)),
-    (Failure(Failure('a')), Failure('a')),
     (Success(Success({})), Success({})),
     (Some(Some(None)), Nothing),
     (Some(Some([])), Some([])),
+
+    # Nope:
+    (Failure(Failure('a')), Failure(Failure('a'))),
 ])
-def test_join(container, merged):
+def test_flatten(container, merged):
     """Ensures that `join` is always returning the correct type."""
-    assert join(container) == merged
+    assert flatten(container) == merged
+
+
+def test_flatten_context():
+    """Ensures that `join` works with Context."""
+    assert flatten(
+        Context.unit(Context.unit(1)),
+    )(Context.Empty) == 1